Rename concurrent module to bodies
